Fix JSCS and tests
