--- app/initializers/offline-globals.js
@@ -1,1 +1,0 @@
-export { default, initialize } from 'ember-flexberry-data/initializers/offline-globals';
