+++ app/models/offline-model.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-data/models/offline-model';
