+++ app/serializers/base.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-data/serializers/base';
