--- app/utils/backup.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-data/utils/backup';
