--- app/utils/generate-unique-id.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-data/utils/generate-unique-id';
