--- app/utils/is-model-instance.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-data/utils/is-model-instance';
