--- app/utils/is-object.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-data/utils/is-object';
