--- app/utils/reload-local-records.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-data/utils/reload-local-records';
